app/requirements.txt
gradio
pandas
numpy
opencv-python-headless
ultralytics
Pillow
